feat: parse profile, add relay column
