+++ src/nostr/event/TextNote.test.ts
@@ -1,10 +1,10 @@
 import assert from 'assert';
 
-import { Kind, type Event as NostrEvent } from 'nostr-tools';
+import { Kind } from 'nostr-tools';
 import { describe, it } from 'vitest';
 
-import TextNote, { MarkedEventTag, markedEventTags } from '@/nostr/event/TextNote';
-import { TagReference } from '@/nostr/parseTextNote';
+import TextNote from '@/nostr/event/TextNote';
+import { MarkedEventTag, markedEventTags } from '@/nostr/event/TextNoteLike';
 
 describe('markedEventTags', () => {
   it('should return an empty array if the event has no tags', () => {
@@ -194,68 +194,6 @@ describe('TextNote', () => {
         relayUrl: null,
       };
       assert.deepStrictEqual(textnote.replyingToEvent(), expected);
-    });
-  });
-
-  describe('#resolveTagReference', () => {
-    it('should resolve a tag reference refers a user', () => {
-      const tagReference: TagReference = {
-        type: 'TagReference',
-        tagIndex: 1,
-        content: '#[1]',
-      };
-      const dummyEvent: NostrEvent = {
-        id: '',
-        sig: '',
-        kind: 1,
-        content: '#[1]',
-        tags: [
-          ['p', '9366708117c4a7edf9178acdce538c95059b9eb3394808cdd90564094172d972'],
-          ['p', '80d3a41d8a00679c0105faac2cdf7643c9ba26835cff096bf7f9c7a0eee8c8fc'],
-        ],
-        created_at: 1678377182,
-        pubkey: '9366708117c4a7edf9178acdce538c95059b9eb3394808cdd90564094172d972',
-      };
-      const textNote = new TextNote(dummyEvent);
-      const result = textNote.resolveTagReference(tagReference);
-      const expected = {
-        type: 'MentionedUser',
-        tagIndex: 1,
-        content: '#[1]',
-        pubkey: '80d3a41d8a00679c0105faac2cdf7643c9ba26835cff096bf7f9c7a0eee8c8fc',
-      };
-
-      assert.deepStrictEqual(result, expected);
-    });
-
-    it('should resolve a tag reference refers an other text note', () => {
-      const tagReference: TagReference = {
-        type: 'TagReference',
-        tagIndex: 1,
-        content: '#[1]',
-      };
-      const dummyEvent: NostrEvent = {
-        id: '',
-        sig: '',
-        kind: 1,
-        content: '',
-        tags: [
-          ['p', '80d3a41d8a00679c0105faac2cdf7643c9ba26835cff096bf7f9c7a0eee8c8fc'],
-          ['e', 'b9cefcb857fa487d5794156e85b30a7f98cb21721040631210262091d86ff6f2', '', 'reply'],
-        ],
-        created_at: 1678377182,
-        pubkey: '9366708117c4a7edf9178acdce538c95059b9eb3394808cdd90564094172d972',
-      };
-      const textNote = new TextNote(dummyEvent);
-      const result = textNote.resolveTagReference(tagReference);
-      const expected = {
-        type: 'MentionedEvent',
-        tagIndex: 1,
-        marker: 'reply',
-        content: '#[1]',
-        eventId: 'b9cefcb857fa487d5794156e85b30a7f98cb21721040631210262091d86ff6f2',
-      };
-      assert.deepStrictEqual(result, expected);
     });
   });
 });
